refactor: 共通UIコンポーネントをshared/componentsに移行（TASK-0103）
Button, Dialog, HeaderUI, FooterUI, SidebarUI, TooltipManager, RewardCardDialogの
7ファイルをpresentation/ui/componentsからshared/componentsに移動し、
旧ファイルを後方互換性用re-exportに変換。
